Adding more instructions
In this commit, I changed the description displayed on screen so the user can provide valid information. For example,
in line 30 I'm asking the user to put the hour in the correct format.

I also changed the database name and the fields used in the SQL Query in the deleteMeeting method because I had incorrect
data.

The last change was ask the users for the amount of meetings that they want to add and execute the gettingValues and
addingToZoomDatabase the times that they want to add different meetings to the Database