Solving the database bug and creating the software GUI
At the line 63 in main.py I changed a little bit of code because I had a bug that didn't allow me to get the password
and meeting ID at certain hours, so the new SQL query is correct. I just had to add an hour interval.

A big change was adding a new python script that creates the GUI for the software application. I did it in order to
have an interface because using the command line is quite annoying. The GUI consists of two tabs right now. The main
tab has a button that take the users to their class meeting. If there is no class, the GUI shows an error message in
a little window.